relocating Card component
