+++ src/components/FeedbackItem.jsx
@@ -1,6 +1,6 @@
 import { FaTimes } from 'react-icons/fa';
 import PropTypes from 'prop-types';
-import Card from './Card';
+import Card from './shared/Card';
 
 function FeedbackItem({ item, handleDelete }) {
   return (
